www: Stop passing a negative uid to gather_receipt_stats_for_user()
uid is defined as an unsigned int, it is also one of the arguments to
gather_receipt_stats_for_user() where the parameter was defined as long
long so we could pass -1 as the uid to indicate we wanted to gather
stats for all users.

This is a bit ugly. So create an enum that contains the possible stats
options; STATS_ALL, STATS_USER and pass one of those in rather than a
negative uid.

So when gathering stats for a specific user, pass in the users uid and
whom as STATS_USER, when gathering stats for *all* users, pass in uid as
0 and whom as STATS_ALL.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>