www/db: Fix a compiler warning in __db_conn()
When compiling with GCC 9.2.1 under Fedora 31 I got the following
compiler warning

  CC	db.o
In file included from /usr/include/string.h:495,
                 from common.h:17,
                 from db.c:19:
In function ‘strncat’,
    inlined from ‘__db_conn’ at db.c:48:3:
/usr/include/bits/string_fortified.h:136:10: warning: ‘__builtin___strncat_chk’ output may be truncated copying 64 bytes from a string of length 64 [-Wstringop-truncation]
  136 |   return __builtin___strncat_chk (__dest, __src, __len, __bos (__dest));
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

This comes from the strncat(3) when creating the database name. The code
is not *technically* wrong, but lets just get rid of the strncat() and
replace it with a memcpy(3) remembering to copy the nul byte.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>