www: Fix various 'warning: unused parameter' compiler warnings
Now that since the previous commit we are compiling with -Wextra,
various new compiler warnings have sprung up. Fix the first batch which
are due to unused function parameters.

Fix this by adding a couple of #define's to common.h which create
short aliases to __attribute__((unused)) which we can use to mark
parameters etc as such.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>