www: Re-work config
This largely re-works the configuration of the app.

Rather than just have a bunch of variables *and* #defines for them, put
all the config variables inside a new struct cfg.

We get rid of receiptomatic_config.h and rename get_config.[ch] ->
config.[ch], config.h contains the struct cfg definition and the default
values for some of the optional config parameters.

We declare struct cfg in receiptomatic-www.c and put an extern
declaration for it in common.h as the config is needed by various bits
of the code.

This work was prompted by the fact that GCC starting with GCC 10 will
default to enabling -fno-common[0] and this will trip up on the fact
that we had global variables declared in a header file that is included
in multiple files which will cause multiple definitions of those
variables.

We could always just compile with -fcommon, but lets take the
opportunity to clean this up and it is a nice clean up...

[0]: https://gcc.gnu.org/gcc-10/porting_to.html#common

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>