Revert "www: audit: Fix a segfault when doing name lookups"
This reverts commit 4404fc530ae5a3036808dd7db1165968c9be9880

This was done manually as there were quite a few conflicts and I've
tried to keep intact the work that was done since.

The above commit was fixing a problem with the log_utmp_host()
threads using the global MYSQL connection handle (via
make_mysql_safe_string().

The way that commit was done was a little convoluted. A much simpler way
is to just declare the connection handle as thread local storage.

So revert the above mentioned commit and then we can do the nicer fix.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>